Resuelto problema Generador ussers polimorfico

+++ Usuarios/Administrador.cs
@@ -16,7 +16,7 @@ namespace Usuarios
         private TerminalConsola terminal;
         private String nombreCompleto;
         private String direccion;
-        private Double numeroDecontacto;
+        private int numeroDecontacto;
         private int idUsuario;
         private int numeroDeComuna;
 
@@ -40,7 +40,7 @@ namespace Usuarios
         public string getDireccion() { return this.direccion; }
 
         public void setNumeroDeContacto(double numero) { this.numeroDecontacto = numero; }
-        public double getNumeroDeContacto() { return this.numeroDecontacto; }
+        public int getNumeroDeContacto() { return this.numeroDecontacto; }
 
         public void setIdUsuario(int numero) { this.idUsuario = numero; }
         public int getIdUsuario() { return this.idUsuario; }
